requirements.txt
# DRApp requirements
#

flask==0.10.1
pymongo==3.2
six==1.10.0
python-dateutil

# Test requirements
unittest2==1.1.0
Flask-Testing==0.4.2